Moved workshop schedule to seperate data file
